COMP: fix

+++ Utilities/VTK/vtkGDCMPolyDataReader.cxx
@@ -345,7 +345,7 @@ int vtkGDCMPolyDataReader::RequestData_RTStructureSetStorage(gdcm::Reader const 
 
       //newPts->SetNumberOfPoints( at.GetNumberOfValues() / 3 );
       //assert( at.GetNumberOfValues() % 3 == 0); // FIXME
-      const float* pts = at.GetValues();
+      const double* pts = at.GetValues();
       vtkIdType buffer[256];
       vtkIdType *ptIds;
       unsigned int npts = at.GetNumberOfValues() / 3;
